+++ src/AllModule/ButtonFunctions.hpp
@@ -1,0 +1,4 @@
+void initializeButton()
+{
+    pinMode(BUTTON_PIN, INPUT_PULLUP);
+}